refactor(utils): fold over-cap git.py into cohesive sibling modules
